+++ contracts/interfaces/IGroth16VerifierP2.sol
@@ -2,9 +2,9 @@ pragma solidity ^0.8.7;
 
 interface IGroth16VerifierP2 {
     function verifyProof(
-        uint[2] calldata _pA,
-        uint[2][2] calldata _pB,
-        uint[2] calldata _pC,
-        uint[2] calldata _pubSignals
+        uint256[2] calldata _pA,
+        uint256[2][2] calldata _pB,
+        uint256[2] calldata _pC,
+        uint256[2] calldata _pubSignals
     ) external view returns (bool);
 }
